support spam progress reports in API logs
